Fix another case of ENV parsing
The broken case was found in one of the failing integration tests.

This required a full re-implementation of the Lexer. I decided to drop
the makeTokenParser since it has been a struggle form the start to manage
the way it munges whitespace.

Instead, implemented the token parsers in the lexer by hand, since we only
needed a few of them.