parser: Handle `ARG` instructions without default
Improve the parser by correctly handling `ARG` instructions that do not
have a default, but do have a `=`. E.g.:
```
ARG FOO=
```
Previously, the parser would consume all characters until EOL as
argument name in this case, which is wrong, since the `=` sign should
not be part of the argument name in this case.