pretty printer: detect escape character from AST
- Add escape character detection logic to the pretty printer
- Replace the escape character on demand for printing
- Add tests ensuring correct default behaviour
- Version bump to 10.0.0 due to API change

When pretty printing, look ahead into the AST to be printed to determine
which escape character to choose. Then use that character during
printing.
This is possibly a major breaking change in the API, because `Arguments
Text` is no longer an instance of `Pretty`. Instead the pretty printer
expects all instructions to be of type `Instruction Text` and not
`Instruction args`. This allows the arguments to the instructions to be
printed with the correct escape character.