heredocs: fix parsing bug with line continuations
Fix parsing error on line continuation.
The `Text.Megaparsec.Lexer.charLiteral` parser will not parse just any
Char, but will instead make exceptions for escaped characters and
quotes. This resulted in a faulty behaviour when encountering a literal
backslash `\` in a heredoc right before a newline.
The `Text.Megaparsec.anySingle` parsed will instead not make exceptions
for escaped characters and therefore parse the content of a heredoc just
as it is in the input, which is the desired behaviour.

Additionally a test case for the `ADD` and `COPY` instruction has been
added to make sure the parser continues to accept this kind of input.

fixes: https://github.com/hadolint/hadolint/issues/699