AST: encode the specific form of pair in the AST
Encode the specific syntax of key/value pair in the AST.

Docker currently supports two forms for key value pairs. The `key=value`
syntax and the deprecated `key value` syntax. These are primarily used
in the `ENV` and `LABEL` instructions.
This change adds encoding for which form was parsed to the syntax tree.
Having this information can be used e.g. to warn about the use of
deprecated syntax and to specify which syntax form the pretty printer
should produce.