syntax: add support for `--link`
- Add support for parsing `--link` flag in `COPY` and `ADD`
  instructions.

With Dockerfile syntax version 1.4, the `--link` flag for the `COPY` and
`ADD` instructions become officially supported. The flag changes the way
the layers are treated during build such that caching can work more
efficient.

See also: https://github.com/moby/buildkit/blob/master/frontend/dockerfile/docs/syntax.md

This change modifies the AST and is not backwards compatible. For `COPY`
and `ADD` instructions, the arguments are now separated from the flags.

This also allows the creation of a default instance for the data
structure containing the flags.

fixes: #81