ast/parser: vairables in port ranges
- Fix parser being unable to parse port ranges when the limits are
  expressed with variables
- Rearrange AST such that the limits of port ranges are ports
- Move tests for `EXPOSE` instructions to their own file

This change introduces a breaking change to the AST.

The old AST was unable to represent port ranges delimited with
variables, since the data type for port ranges mandated that a limit
must be expressed as an integer. This creates problems when limits of a
range are expressed as variables (e.g. as ARGs). The change to the AST
defines the limits of a port range as port. This has the advantage that
any way a port can be expressed can also be used to express a port range
limit, including variables.

required-for: https://github.com/hadolint/hadolint/issues/876