fix: parse fractional arguments to flags
Parse fractional numbers as arguments to flags for the `HEALTHCHECK`
instruction.

The `HEALTHCHECK` instruction can take an `--interval`, `--timeout` or
`--start-period` flag specifying details to the healthcheck repetition
behaviour. This can also be specified with fractions (of secods, minutes
or hours), e.g.:

```Dockerfile
HEALTHCHECK \
  --interval=0.5s \
  CMD curl -f http://localhost
```

With the caveat that somewhere from parsing to conversion to a DiffTime
some precision is lost (DiffTime has a precision of 10^-12s), this
change adds support for the arguments to these interval flags to be
specified as fractional numbers.

fixes: https://github.com/hadolint/hadolint/issues/893
Signed-off-by: Moritz Röhrich <moritz@ildefons.de>